refactor: extend res.{end,send,json}, redirect
